agents/nutrition_expert/requirements.txt
google-adk
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx